test: suppress make flags from environment
