+++ pytest_jobserver/filesystem.py
@@ -1,6 +1,6 @@
 import os
 import stat
-from typing import NewType, Tuple
+from typing import Iterable, NewType, Optional, Tuple
 
 FileDescriptor = NewType("FileDescriptor", int)
 # A 2-tuple of file descriptors, representing a read/write pair
@@ -13,3 +13,11 @@ def is_fifo(path: str) -> bool:
 
 def is_rw_ok(path: str) -> bool:
     return os.access(path, os.R_OK | os.W_OK)
+
+
+def first_environ_get(keys: Iterable[str]) -> Optional[str]:
+    for key in keys:
+        value = os.environ.get(key)
+        if value is not None:
+            return value
+    return None
